requirements.txt
svglib
reportlab
pillow
pycairo
rlPyCairo

# Evaluation dependencies
rich>=13.0.0
rich-click>=1.7.0
aiohttp>=3.8.0
pydantic>=2.0.0
tqdm>=4.64.0
transformers>=4.30.0
datasets>=2.14.0